checkpoint1 of rewriting game 'engine'
