+++ ml/src/data_structures/__init__.py
@@ -1,0 +1,7 @@
+from .datapoint import DataPoint
+from .round_data import MoveData, RoundData
+
+__all__ = [
+    "DataPoint",
+    "MoveData", "RoundData",
+]
